pyproject.toml
[project]
name = "mriynyk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.15.0",
    "pandas>=2.3.3",
    "pgvector>=0.4.1",
    "psycopg[binary]>=3.2.10",
    "python-dotenv>=1.1.1",
    "pyarrow>=22.0.0",
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
]
